+++ driving_log_replayer_analyzer/driving_log_replayer_analyzer/plot/scatter_plot.py
@@ -16,7 +16,7 @@ from pathlib import Path
 
 import plotly.express as px
 
-from .plot_base import PlotBase
+from . import PlotBase
 
 
 class ScatterPlot(PlotBase):
